fix: finalizei a implementação de Fragments para a criação de conta
