+++ app/src/main/res/layout/activity_criar_conta.xml
@@ -10,7 +10,6 @@
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:fitsSystemWindows="true"
         android:theme="@style/Theme.DenTeeth.AppBarOverlay">
 
         <com.google.android.material.appbar.MaterialToolbar
@@ -20,11 +19,11 @@
 
     </com.google.android.material.appbar.AppBarLayout>
 
-    <com.google.android.material.progressindicator.LinearProgressIndicator
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_anchor="@id/tbNavbar"
-        app:layout_anchorGravity="bottom|center" />
+<!--    <com.google.android.material.progressindicator.LinearProgressIndicator-->
+<!--        android:layout_width="match_parent"-->
+<!--        android:layout_height="wrap_content"-->
+<!--        app:layout_anchor="@id/tbNavbar"-->
+<!--        app:layout_anchorGravity="bottom|center" />-->
 
     <include layout="@layout/content_criar_conta" />
