+++ app/src/main/res/layout/content_criar_conta.xml
@@ -7,7 +7,7 @@
     android:padding="8dp"
     app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
-    <androidx.fragment.app.FragmentContainerView
+    <fragment
         android:id="@+id/nav_host_fragment_content_criar_conta"
         android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="0dp"
